route and controller for get user neck angle records
