+++ .env
@@ -1,7 +1,1 @@
-PORT=5000
-# MONGODB_URI=mongodb://cervitech:bD1Vz7ATLc2gCbEL@cluster0.kw0de01.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
-# MONGODB_URI=mongodb+srv://cervitech:bD1Vz7ATLc2gCbEL@cluster0.zmtbk6a.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0
-MONGODB_URI=mongodb+srv://cervitech_devs:nHMLPU9.!EGwdtX@cervitechdb.tzqmaqb.mongodb.net/?retryWrites=true&w=majority
-REDIS_URL=redis://localhost:6379
-JWT_SECRET=temp_jwt
-SENDGRID_API_KEY=your_sendgrid_api_key_here
+
