+++ src/controllers/user.controller.ts
@@ -3,10 +3,10 @@ import { Request, Response } from "express";
 import { PictureUrlUpdateViewModel } from "../viewmodels/PictureUrlUpdateViewModel";
 import { AppUserService } from "../services/appUserServices/appUserService.service";
 import { GetApiResponseMessages, ApiResponseStatus } from "../helpers/apiResponse";
-import { DataResult } from "../helpers/dataResult";
+import { DataResult } from "../helpers/DataResult";
 import { UpdateUserRequest } from "../types/user.types";
 import { AuthenticatedRequest } from "../middlewares/auth.middleware";
-import { CustomException } from "../helpers/customException";
+import { CustomException } from "../helpers/CustomException";
 import { logger } from "../utils/logger";
 
 export class UserController {
@@ -301,9 +301,5 @@ export class UserController {
       res.status(400).json({ error: err.message });
     }
   };
-
-  
-
-
 
 }
