--- src/viewmodels/AbbreviatedNeckAngleRecordViewModel.ts
@@ -1,5 +1,0 @@
-export interface AbbreviatedNeckAngleRecordViewModel {
-  angle: number;
-  craniumVertebralAngle: number;
-  dateTimeRecorded: Date;
-}
